Split user button into it's own component
