+++ packages/studio-base/src/components/AppBar/UserMenu.tsx
@@ -2,95 +2,42 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import PersonIcon from "@mui/icons-material/Person";
 import {
-  Avatar,
   Divider,
-  IconButton,
-  IconButtonProps,
   ListItemText,
   Menu,
   MenuItem,
   PopoverPosition,
   PopoverReference,
-  Tooltip,
   Typography,
 } from "@mui/material";
 import { useSnackbar } from "notistack";
-import { forwardRef, useCallback } from "react";
-import tinycolor from "tinycolor2";
+import { SetStateAction, useCallback } from "react";
 import { makeStyles } from "tss-react/mui";
 
 import Logger from "@foxglove/log";
-import {
-  APP_BAR_FOREGROUND_COLOR,
-  APP_BAR_PRIMARY_COLOR,
-} from "@foxglove/studio-base/components/AppBar/constants";
-import { useCurrentUser, User } from "@foxglove/studio-base/context/CurrentUserContext";
+import { useCurrentUser, User, UserType } from "@foxglove/studio-base/context/CurrentUserContext";
 import { useConfirm } from "@foxglove/studio-base/hooks/useConfirm";
 
 const log = Logger.getLogger(__filename);
 
-const useStyles = makeStyles()((theme) => ({
-  tooltip: {
-    marginTop: `${theme.spacing(0.5)} !important`,
-  },
-  avatar: {
-    color: theme.palette.common.white,
-    backgroundColor: APP_BAR_PRIMARY_COLOR,
-    height: theme.spacing(3.5),
-    width: theme.spacing(3.5),
-  },
-  iconButton: {
-    padding: theme.spacing(1),
-    borderRadius: 0,
-
-    "&:hover": {
-      backgroundColor: tinycolor(APP_BAR_FOREGROUND_COLOR).setAlpha(0.08).toRgbString(),
-    },
-    "&.Mui-selected": {
-      backgroundColor: APP_BAR_PRIMARY_COLOR,
-    },
-  },
+const useStyles = makeStyles()({
   menuList: {
     minWidth: 200,
   },
-  userIconImage: {
-    objectFit: "cover",
-    width: "100%",
-  },
-}));
-
-type UserIconProps = IconButtonProps & {
-  currentUser?: User;
-};
-
-export const UserIconButton = forwardRef<HTMLButtonElement, UserIconProps>((props, ref) => {
-  const { classes } = useStyles();
-  const { currentUser: me, title, ...otherProps } = props;
-
-  return (
-    <Tooltip
-      classes={{ tooltip: classes.tooltip }}
-      title={me?.email ?? title ?? "Profile"}
-      arrow={false}
-    >
-      <IconButton {...otherProps} ref={ref} className={classes.iconButton}>
-        <Avatar className={classes.avatar} variant="rounded">
-          {me?.avatarImageUrl != undefined && (
-            <img
-              src={me.avatarImageUrl}
-              referrerPolicy="same-origin"
-              className={classes.userIconImage}
-            />
-          )}
-          {me?.avatarImageUrl == undefined && <PersonIcon />}
-        </Avatar>
-      </IconButton>
-    </Tooltip>
-  );
 });
-UserIconButton.displayName = "UserIconButton";
+
+export type UserButtonProps = {
+  disableSignIn?: boolean;
+  currentUser?: User;
+  currentUserType?: UserType;
+  signIn?: () => void;
+  userMenuOpen: boolean;
+  setUserAnchorEl: (value: SetStateAction<HTMLElement | undefined>) => void;
+  prefsDialogOpen: boolean;
+  // eslint-disable-next-line @foxglove/no-boolean-parameters
+  setPrefsDialogOpen: (open: boolean) => void;
+};
 
 export function UserMenu({
   anchorEl,
